Fixed weird bug when a file is invalid and it is included twice.
It was the case when the compiled file was not expired, but its content were bogus (corrupted file, it was being compiled by another process, the template has changed its path). Because we were creating lamba on those cases, in a second `Haanga::load` it would attempt to `require` the bogus compiled template again (instead of using the previous compiled lamba).

This fix removes the lamba generation, instead uses eval() to create a global function. It also invalidates the compiled file.